requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
diskcache>=5.6.0
tenacity>=8.2.0
